ref: Rename SupabaseIntegrationCredentialsStore to IntegrationCredentialsStore
Rename autogpt_libs.supabase_integration_credentials_store to integration_credentials_stroe
Rename SupabaseIntegrationCredentialsStore to IntegrationCredentialsStore
