+++ autogpt_platform/autogpt_libs/autogpt_libs/integration_credentials_store/__init__.py
@@ -1,8 +1,8 @@
-from .store import SupabaseIntegrationCredentialsStore
+from .store import IntegrationCredentialsStore
 from .types import Credentials, APIKeyCredentials, OAuth2Credentials
 
 __all__ = [
-    "SupabaseIntegrationCredentialsStore",
+    "IntegrationCredentialsStore",
     "Credentials",
     "APIKeyCredentials",
     "OAuth2Credentials",
