+++ autogpt_platform/autogpt_libs/autogpt_libs/integration_credentials_store/types.py
