+++ autogpt_platform/backend/backend/blocks/ai_music_generator.py
@@ -4,7 +4,7 @@ from enum import Enum
 from typing import Literal
 
 import replicate
-from autogpt_libs.supabase_integration_credentials_store.types import APIKeyCredentials
+from autogpt_libs.integration_credentials_store.types import APIKeyCredentials
 from pydantic import SecretStr
 
 from backend.data.block import Block, BlockCategory, BlockOutput, BlockSchema
