+++ autogpt_platform/backend/backend/blocks/ai_shortform_video_block.py
@@ -3,7 +3,7 @@ import time
 from enum import Enum
 from typing import Literal
 
-from autogpt_libs.supabase_integration_credentials_store.types import APIKeyCredentials
+from autogpt_libs.integration_credentials_store.types import APIKeyCredentials
 from pydantic import SecretStr
 
 from backend.data.block import Block, BlockCategory, BlockOutput, BlockSchema
