+++ autogpt_platform/backend/backend/blocks/google_maps.py
@@ -1,7 +1,7 @@
 from typing import Literal
 
 import googlemaps
-from autogpt_libs.supabase_integration_credentials_store.types import APIKeyCredentials
+from autogpt_libs.integration_credentials_store.types import APIKeyCredentials
 from pydantic import BaseModel, SecretStr
 
 from backend.data.block import Block, BlockCategory, BlockOutput, BlockSchema
