+++ autogpt_platform/backend/backend/blocks/ideogram.py
@@ -1,7 +1,7 @@
 from enum import Enum
 from typing import Any, Dict, Literal, Optional
 
-from autogpt_libs.supabase_integration_credentials_store.types import APIKeyCredentials
+from autogpt_libs.integration_credentials_store.types import APIKeyCredentials
 from pydantic import SecretStr
 from requests.exceptions import RequestException
