+++ autogpt_platform/backend/backend/blocks/jina/_auth.py
@@ -1,6 +1,6 @@
 from typing import Literal
 
-from autogpt_libs.supabase_integration_credentials_store.types import APIKeyCredentials
+from autogpt_libs.integration_credentials_store.types import APIKeyCredentials
 from pydantic import SecretStr
 
 from backend.data.model import CredentialsField, CredentialsMetaInput
