+++ autogpt_platform/backend/backend/blocks/medium.py
@@ -1,7 +1,7 @@
 from enum import Enum
 from typing import List, Literal
 
-from autogpt_libs.supabase_integration_credentials_store.types import APIKeyCredentials
+from autogpt_libs.integration_credentials_store.types import APIKeyCredentials
 from pydantic import SecretStr
 
 from backend.data.block import Block, BlockCategory, BlockOutput, BlockSchema
