+++ autogpt_platform/backend/backend/blocks/replicate_flux_advanced.py
@@ -3,7 +3,7 @@ from enum import Enum
 from typing import Literal
 
 import replicate
-from autogpt_libs.supabase_integration_credentials_store.types import APIKeyCredentials
+from autogpt_libs.integration_credentials_store.types import APIKeyCredentials
 from pydantic import SecretStr
 from replicate.helpers import FileOutput
