+++ autogpt_platform/backend/backend/blocks/text_to_speech_block.py
@@ -1,6 +1,6 @@
 from typing import Any, Literal
 
-from autogpt_libs.supabase_integration_credentials_store.types import APIKeyCredentials
+from autogpt_libs.integration_credentials_store.types import APIKeyCredentials
 from pydantic import SecretStr
 
 from backend.data.block import Block, BlockCategory, BlockOutput, BlockSchema
