+++ autogpt_platform/backend/backend/data/block.py
@@ -15,7 +15,7 @@ from typing import (
 
 import jsonref
 import jsonschema
-from autogpt_libs.supabase_integration_credentials_store.types import Credentials
+from autogpt_libs.integration_credentials_store.types import Credentials
 from prisma.models import AgentBlock
 from pydantic import BaseModel
